feat(timer): add style to delete confirmation dialog
