+++ lib/timer/view/timer_page.dart
@@ -5,7 +5,7 @@ import 'package:kubrs_app/scramble/bloc/scramble_bloc.dart';
 import 'package:kubrs_app/solve/bloc/solve_bloc.dart';
 import 'package:kubrs_app/solve/repository/solve_repository.dart';
 import 'package:kubrs_app/timer/bloc/timer_bloc.dart';
-import 'package:kubrs_app/timer/view/button/delete_button.dart';
+import 'package:kubrs_app/timer/view/button/delete_solve_button.dart';
 import 'package:kubrs_app/timer/view/button/tag/dnf_tag_toggle_button.dart';
 import 'package:kubrs_app/timer/view/button/tag/plus_two_tag_toggle_button.dart';
 import 'package:kubrs_app/timer/view/timer_gesture_detector.dart';
@@ -86,7 +86,7 @@ class TimerView extends StatelessWidget {
               child: Row(
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: [
-                  DeleteButton(),
+                  DeleteSolveButton(),
                   PlusTwoTagToggleButton(),
                   DNFTagToggleButton(),
                 ],
